backend/requirements.txt
asgiref==3.7.2
black==23.3.0
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
Django==4.2.2
django-cors-headers==4.1.0
django-stubs==4.2.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
environs==9.5.0
flake8==6.0.0
isort==5.12.0
mypy==1.4.1
psycopg2-binary==2.9.6
pytest-django==4.5.2
shortuuid==1.0.9
sqlparse==0.4.4
